Usuarios admin agregados, cambio en la respuestas de los controller

+++ routes/api.php
@@ -14,6 +14,7 @@ use App\Http\Controllers\ClienteController as Cliente;
 use App\Http\Controllers\DireccionClienteController as DireccionCliente;
 use App\Http\Controllers\SucursalController as Sucursal;
 use App\Http\Controllers\UsuarioController as Usuario;
+use App\Http\Controllers\AdminController as Admin;
 
 /*
 |--------------------------------------------------------------------------
@@ -33,47 +34,55 @@ Route::post('/api/login', [Usuario::class, 'login']);
 
 Route::post('/api/logout', [Usuario::class, 'logout'])->middleware(['auth:sanctum']);
 ###########################################
+########    RUTAS DE ADMIN   ###########
+###########################################
+Route::post('api/admin', [Admin::class, 'crearAdmin']);
+
+Route::put('api/admin/{id}', [Admin::class, 'actualizarAdmin']);
+
+Route::delete('api/admin/{id}', [Admin::class, 'borrarAdmin'])->middleware(['auth:sanctum', 'ability:admin']);
+###########################################
 ########    RUTAS DE EMPRESAS   ###########
 ###########################################
-Route::get('/api/empresas',  [Empresa::class, 'getAllEmpresas'])->middleware(['auth:sanctum', 'ability:empresa, admin']);
+Route::get('/api/empresas',  [Empresa::class, 'getAllEmpresas'])->middleware(['auth:sanctum', 'ability:admin,admin_clientes']);
 
 Route::get('/api/empresas/{id}',  [Empresa::class, 'getEmpresa']);
 
 Route::post('/api/empresas',  [Empresa::class, 'guardarEmpresa']);
 
-Route::put('/api/empresas/{id}',  [Empresa::class, 'actualizarEmpresa'])->middleware(['auth:sanctum', 'ability:empresa']);
+Route::put('/api/empresas/{id}',  [Empresa::class, 'actualizarEmpresa'])->middleware(['auth:sanctum', 'ability:empresa,admin,admin_clientes']);
 
-Route::delete('/api/empresas/{id}',  [Empresa::class, 'eliminarEmpresa'])->middleware(['auth:sanctum', 'ability:empresa']);
+Route::delete('/api/empresas/{id}',  [Empresa::class, 'eliminarEmpresa'])->middleware(['auth:sanctum', 'ability:empresa,admin,admin_clientes']);
 ###########################################
 ########    RUTAS DE SUCURSALES   ###########
 ###########################################
 Route::get('/api/sucursales/{id}', [Sucursal::class, 'getSucursales']);
-Route::post('/api/sucursales/{id}', [Sucursal::class, 'guardarSucursal'])->middleware(['auth:sanctum', 'ability:empresa, admin']);
-Route::put('/api/sucursales/{empresa_id}/{direccion_id}', [Sucursal::class, 'actualizarSucursal'])->middleware(['auth:sanctum', 'ability:empresa, admin']);
-Route::delete('/api/sucursales/{empresa_id}/{direccion_id}', [Sucursal::class, 'eliminarSucursal'])->middleware(['auth:sanctum', 'ability:empresa, admin']);
+Route::post('/api/sucursales/{id}', [Sucursal::class, 'guardarSucursal'])->middleware(['auth:sanctum', 'ability:empresa,admin,admin_clientes']);
+Route::put('/api/sucursales/{empresa_id}/{direccion_id}', [Sucursal::class, 'actualizarSucursal'])->middleware(['auth:sanctum', 'ability:empresa,admin,admin_clientes']);
+Route::delete('/api/sucursales/{empresa_id}/{direccion_id}', [Sucursal::class, 'eliminarSucursal'])->middleware(['auth:sanctum', 'ability:empresa,admin,admin_clientes']);
 ###########################################
 ########    RUTAS DE SOLICITUDES   ########
 ###########################################
-Route::get('/api/solicitudes',  [Solicitudes::class, 'getAllSolicitudes'])->middleware(['auth:sanctum', 'ability:admin']);
+Route::get('/api/solicitudes',  [Solicitudes::class, 'getAllSolicitudes'])->middleware(['auth:sanctum', 'ability:admin,admin_clientes']);
 
-Route::put('/api/solicitudes/{id}',  [Solicitudes::class, 'actualizarSolicitud'])->middleware(['auth:sanctum', 'ability:admin']);
+Route::put('/api/solicitudes/{id}',  [Solicitudes::class, 'actualizarSolicitud'])->middleware(['auth:sanctum', 'ability:admin,admin_clientes']);
 
 //se rechaza la empresa y se elimina xd
-Route::delete('/api/solicitudes/{id}',  [Solicitudes::class, 'rechazarSolicitud'])->middleware(['auth:sanctum', 'ability:admin']);
+Route::delete('/api/solicitudes/{id}',  [Solicitudes::class, 'rechazarSolicitud'])->middleware(['auth:sanctum', 'ability:admin,admin_clientes']);
 
 ###########################################
 ########    RUTAS DE Clientes    ##########
 ###########################################
 
-Route::get('/api/clientes',  [Cliente::class, 'getAllClientes'])->middleware(['auth:sanctum', 'ability:empresa, admin']);
+Route::get('/api/clientes',  [Cliente::class, 'getAllClientes'])->middleware(['auth:sanctum', 'ability:empresa,admin,admin_clientes']);
 
 Route::get('/api/clientes/{id}',  [Cliente::class, 'getCliente']);
 
 Route::post('/api/clientes',  [Cliente::class, 'guardarCliente'])->middleware(['auth:sanctum', 'ability:empresa']);
 
-Route::put('/api/clientes/{id}',  [Cliente::class, 'actualizarCliente'])->middleware(['auth:sanctum', 'ability:admin,cliente,empresa']);
+Route::put('/api/clientes/{id}',  [Cliente::class, 'actualizarCliente'])->middleware(['auth:sanctum', 'ability:admin,cliente,empresa,admin_clientes']);
 
-Route::delete('/api/clientes/{id}',  [Cliente::class, 'eliminarCliente'])->middleware(['auth:sanctum', 'ability:cliente,empresa']);
+Route::delete('/api/clientes/{id}',  [Cliente::class, 'eliminarCliente'])->middleware(['auth:sanctum', 'ability:cliente,empresa,admin_clientes,admin']);
 
 ###########################################
 ##    RUTAS DE Direccion Clientes    ######
@@ -81,11 +90,11 @@ Route::delete('/api/clientes/{id}',  [Cliente::class, 'eliminarCliente'])->middl
 
 Route::get('/api/clientes/{id}/direcciones',  [DireccionCliente::class, 'getClienteDirecciones']);
 
-Route::post('/api/clientes/{id}/direcciones',  [DireccionCliente::class, 'guardarClienteDireccion'])->middleware(['auth:sanctum', 'ability:cliente']);
+Route::post('/api/clientes/{id}/direcciones',  [DireccionCliente::class, 'guardarClienteDireccion'])->middleware(['auth:sanctum', 'ability:cliente,admin_clientes,admin']);
 
-Route::put('/api/clientes/{id}/direcciones/{id_direccion}',  [DireccionCliente::class, 'actualizarClienteDireccion'])->middleware(['auth:sanctum', 'ability:cliente']);
+Route::put('/api/clientes/{id}/direcciones/{id_direccion}',  [DireccionCliente::class, 'actualizarClienteDireccion'])->middleware(['auth:sanctum', 'ability:cliente,admin_clientes,admin']);
 
-Route::delete('/api/clientes/{id}/direcciones/{id_direccion}',  [DireccionCliente::class, 'eliminarClienteDireccion'])->middleware(['auth:sanctum', 'ability:cliente']);
+Route::delete('/api/clientes/{id}/direcciones/{id_direccion}',  [DireccionCliente::class, 'eliminarClienteDireccion'])->middleware(['auth:sanctum', 'ability:cliente,admin_clientes,admin']);
 #RUTAS MODULO 1
 Route::get('/api/inventario', [InventarioController::class, 'verInventario']);
 Route::post('/api/inventario',[InventarioController::class, 'buscarProductos']);
